+++ ext/JEMExt/JEMExt.jl
@@ -5,5 +5,6 @@ using CounterfactualExplanations.Models: Models
 using JointEnergyModels: JointEnergyModels
 
 include("jem.jl")
+include("pretrained.jl")
 
 end