+++ src/deprecated.jl
@@ -75,3 +75,7 @@ Base.@deprecate FluxEnsemble(data::CounterfactualData) DeepEnsemble()(
 )
 
 Base.@deprecate train!(vae, X, y) GenerativeModels.train!(vae, X)
+
+Base.@deprecate load_mnist_mlp() Models.load_mnist_model(MLP())
+
+Base.@deprecate load_mnist_ensemble() Models.load_mnist_model(DeepEnsemble())